Extract medical history editor
