+++ js/context-cards.js
@@ -1,7 +1,7 @@
 // context-cards.js — 9 context card editors, summaries, health dots, interpretive lens
 
 import { state } from './state.js';
-import { COMMON_CONDITIONS, DIET_TYPES, DIET_RESTRICTIONS, DIET_PATTERNS, BOWEL_FREQUENCY, STOOL_CONSISTENCY, BLOATING_SEVERITY, GAS_SEVERITY, ACID_REFLUX, BURPING, NAUSEA, APPETITE, ABDOMINAL_PAIN, FOOD_SENSITIVITIES, EXERCISE_FREQ, EXERCISE_TYPES, EXERCISE_INTENSITY, DAILY_MOVEMENT, SLEEP_DURATIONS, SLEEP_QUALITY, SLEEP_SCHEDULE, SLEEP_ROOM_TEMP, SLEEP_ISSUES, SLEEP_ENVIRONMENT, SLEEP_PRACTICES, LIGHT_AM, LIGHT_DAYTIME, LIGHT_UV, LIGHT_EVENING, LIGHT_COLD, LIGHT_GROUNDING, LIGHT_SCREEN_TIME, LIGHT_TECH_ENV, LIGHT_MEAL_TIMING, STRESS_LEVELS, STRESS_SOURCES, STRESS_MGMT, LOVE_STATUS, LOVE_SATISFACTION, LOVE_LIBIDO, LOVE_FREQUENCY, LOVE_ORGASM, LOVE_RELATIONSHIP, LOVE_CONCERNS, ENV_SETTING, ENV_CLIMATE, ENV_WATER, ENV_WATER_CONCERNS, ENV_EMF, ENV_EMF_MITIGATION, ENV_HOME_LIGHT, ENV_AIR, ENV_TOXINS, ENV_BUILDING } from './constants.js';
+import { DIET_TYPES, DIET_RESTRICTIONS, DIET_PATTERNS, BOWEL_FREQUENCY, STOOL_CONSISTENCY, BLOATING_SEVERITY, GAS_SEVERITY, ACID_REFLUX, BURPING, NAUSEA, APPETITE, ABDOMINAL_PAIN, FOOD_SENSITIVITIES, EXERCISE_FREQ, EXERCISE_TYPES, EXERCISE_INTENSITY, DAILY_MOVEMENT, SLEEP_DURATIONS, SLEEP_QUALITY, SLEEP_SCHEDULE, SLEEP_ROOM_TEMP, SLEEP_ISSUES, SLEEP_ENVIRONMENT, SLEEP_PRACTICES, LIGHT_AM, LIGHT_DAYTIME, LIGHT_UV, LIGHT_EVENING, LIGHT_COLD, LIGHT_GROUNDING, LIGHT_SCREEN_TIME, LIGHT_TECH_ENV, LIGHT_MEAL_TIMING, STRESS_LEVELS, STRESS_SOURCES, STRESS_MGMT, LOVE_STATUS, LOVE_SATISFACTION, LOVE_LIBIDO, LOVE_FREQUENCY, LOVE_ORGASM, LOVE_RELATIONSHIP, LOVE_CONCERNS, ENV_SETTING, ENV_CLIMATE, ENV_WATER, ENV_WATER_CONCERNS, ENV_EMF, ENV_EMF_MITIGATION, ENV_HOME_LIGHT, ENV_AIR, ENV_TOXINS, ENV_BUILDING } from './constants.js';
 import { escapeHTML, hashString, showNotification, hasCardContent } from './utils.js';
 import { formatTime, getTimeFormat, parseTimeInput } from './theme.js';
 import { saveImportedData, getActiveData } from './data.js';
@@ -44,6 +44,24 @@ import {
   renderNoteField,
   contextEditorActions,
 } from './context-card-editor-ui.js';
+import {
+  configureMedicalHistoryEditor,
+  openDiagnosesEditor,
+  renderDiagnosesModal,
+  filterConditionSuggestions,
+  selectConditionSuggestion,
+  closeSuggestionsOnClickOutside,
+  syncDiagnosesNote,
+  addCondition,
+  deleteCondition,
+  addFamilyHistoryEntry,
+  deleteFamilyHistoryEntry,
+  filterFamilyConditionSuggestions,
+  selectFamilyConditionSuggestion,
+  saveDiagnoses,
+  closeDiagnoses,
+  clearDiagnoses,
+} from './context-card-medical-history-editor.js';
 
 export {
   getConditionsSummary,
@@ -68,6 +86,23 @@ export {
   renderNoteField,
   contextEditorActions,
 } from './context-card-editor-ui.js';
+export {
+  openDiagnosesEditor,
+  renderDiagnosesModal,
+  filterConditionSuggestions,
+  selectConditionSuggestion,
+  closeSuggestionsOnClickOutside,
+  syncDiagnosesNote,
+  addCondition,
+  deleteCondition,
+  addFamilyHistoryEntry,
+  deleteFamilyHistoryEntry,
+  filterFamilyConditionSuggestions,
+  selectFamilyConditionSuggestion,
+  saveDiagnoses,
+  closeDiagnoses,
+  clearDiagnoses,
+} from './context-card-medical-history-editor.js';
 
 function _renderDietContaminants() {
   const warnings = scanDietForContaminants(state.importedData.diet);
@@ -340,310 +375,10 @@ export function saveAndRefresh(msg, field) {
   loadContextHealthDots();
 }
 
+configureMedicalHistoryEditor({ recordChange, saveAndRefresh });
+
 function getTimePlaceholder() {
   return getTimeFormat() === '24h' ? 'HH:MM' : 'H:MM AM';
-}
-
-// ═══════════════════════════════════════════════
-// DIAGNOSES / MEDICAL CONDITIONS
-// ═══════════════════════════════════════════════
-
-export function openDiagnosesEditor() {
-  const modal = document.getElementById("detail-modal");
-  const overlay = document.getElementById("modal-overlay");
-  const current = state.importedData.diagnoses || { conditions: [], note: '' };
-  renderDiagnosesModal(modal, current);
-  overlay.classList.add("show");
-}
-
-// Relatives surfaced in the Family History subsection. First-degree
-// (mother/father/sibling/child) plus grandparents — covers the bulk of
-// clinically-relevant heritable risk without sprawling into the
-// "aunt/uncle/cousin" tail where signal-to-noise drops fast.
-const FAMILY_RELATIVES = [
-  { key: 'mother',                 label: 'Mother' },
-  { key: 'father',                 label: 'Father' },
-  { key: 'sibling',                label: 'Sibling' },
-  { key: 'child',                  label: 'Child' },
-  { key: 'maternal_grandmother',   label: 'Maternal grandmother' },
-  { key: 'maternal_grandfather',   label: 'Maternal grandfather' },
-  { key: 'paternal_grandmother',   label: 'Paternal grandmother' },
-  { key: 'paternal_grandfather',   label: 'Paternal grandfather' },
-];
-function _relativeLabel(key) {
-  return FAMILY_RELATIVES.find(r => r.key === key)?.label || key;
-}
-
-export function renderDiagnosesModal(modal, current) {
-  const conditions = current.conditions || [];
-  const familyHistory = Array.isArray(current.familyHistory) ? current.familyHistory : [];
-  let html = '';
-  if (conditions.length > 0) {
-    html += `<div class="ctx-conditions-list" id="ctx-conditions-list">`;
-    for (let i = 0; i < conditions.length; i++) {
-      const c = conditions[i];
-      html += `<div class="ctx-condition-item">
-        <span class="ctx-condition-name">${escapeHTML(c.name)}</span>
-        ${c.severity ? `<span class="goals-severity-badge severity-${c.severity}">${c.severity}</span>` : ''}
-        ${c.since ? `<span class="ctx-condition-since">since ${escapeHTML(c.since)}</span>` : ''}
-        <button class="goals-delete-btn" onclick="deleteCondition(${i})">&times;</button>
-      </div>`;
-    }
-    html += `</div>`;
-  }
-  html += `<div class="ctx-field-group"><label class="ctx-field-label">Add condition</label>
-    <div class="ctx-add-condition">
-      <div class="ctx-autocomplete-wrapper">
-        <input type="text" class="ctx-note-input" id="condition-input" placeholder="Type condition name..." oninput="filterConditionSuggestions()" onfocus="filterConditionSuggestions()">
-        <div class="ctx-suggestions" id="condition-suggestions"></div>
-      </div>
-      <input type="text" class="ctx-note-input" id="condition-since" placeholder="Since (e.g. 2020)" style="width:100px">
-      <button class="import-btn import-btn-primary" onclick="addCondition()">Add</button>
-    </div>
-    <div class="ctx-btn-group" id="condition-severity" style="margin-top:8px">
-      <button type="button" class="ctx-btn-option active" onclick="selectCtxOption(this,'condition-severity')">major</button>
-      <button type="button" class="ctx-btn-option" onclick="selectCtxOption(this,'condition-severity')">mild</button>
-      <button type="button" class="ctx-btn-option" onclick="selectCtxOption(this,'condition-severity')">minor</button>
-    </div>
-  </div>`;
-  // ── Family History subsection ──
-  // Relative-key → emoji for the entry-row chip. Subtle visual hook so users
-  // can scan a long family-history list by "who" before "what".
-  const RELATIVE_EMOJI = {
-    mother: '👩', father: '👨', sibling: '👫', child: '🧒',
-    maternal_grandmother: '👵', maternal_grandfather: '👴',
-    paternal_grandmother: '👵', paternal_grandfather: '👴',
-  };
-  html += `<div class="ctx-family-history" id="ctx-family-section">
-    <div class="ctx-family-head">
-      <label class="ctx-field-label">Family history</label>
-      <span class="ctx-family-count">${familyHistory.length || ''}</span>
-    </div>
-    <div class="ctx-modal-hint">Genetic + environmental signal. Affects risk interpretation — e.g. a father's MI at 52 reframes a borderline LDL.</div>`;
-  if (familyHistory.length > 0) {
-    // Sort for display: same order as FAMILY_RELATIVES, then preserve add order.
-    const relOrder = new Map(FAMILY_RELATIVES.map((r, i) => [r.key, i]));
-    const indexed = familyHistory.map((e, i) => ({ e, i }));
-    indexed.sort((a, b) => (relOrder.get(a.e.relative) ?? 99) - (relOrder.get(b.e.relative) ?? 99));
-    html += `<div class="ctx-family-list" id="ctx-family-list">`;
-    for (const { e, i } of indexed) {
-      const emoji = RELATIVE_EMOJI[e.relative] || '👤';
-      html += `<div class="ctx-family-item">
-        <span class="ctx-family-relative" title="${escapeHTML(_relativeLabel(e.relative))}">${emoji} <span class="ctx-family-relative-label">${escapeHTML(_relativeLabel(e.relative))}</span></span>
-        <span class="ctx-family-condition">${escapeHTML(e.condition || '')}</span>
-        ${e.onsetAge != null && e.onsetAge !== '' ? `<span class="ctx-family-age">age ${escapeHTML(String(e.onsetAge))}</span>` : ''}
-        ${e.note ? `<span class="ctx-family-note">${escapeHTML(e.note)}</span>` : ''}
-        <button class="goals-delete-btn" onclick="deleteFamilyHistoryEntry(${i})" aria-label="Remove entry">&times;</button>
-      </div>`;
-    }
-    html += `</div>`;
-  }
-  // Add form: two stacked rows so 4 inputs + button don't fight on one line.
-  // Row 1: who + what. Row 2: when + optional context + the Add CTA.
-  html += `<div class="ctx-family-add">
-    <div class="ctx-family-add-row">
-      <select class="ctx-note-input ctx-family-select" id="fh-relative" aria-label="Relative">
-        <optgroup label="Parents">
-          <option value="mother">Mother</option>
-          <option value="father">Father</option>
-        </optgroup>
-        <optgroup label="Siblings & Children">
-          <option value="sibling">Sibling</option>
-          <option value="child">Child</option>
-        </optgroup>
-        <optgroup label="Maternal grandparents">
-          <option value="maternal_grandmother">Maternal grandmother</option>
-          <option value="maternal_grandfather">Maternal grandfather</option>
-        </optgroup>
-        <optgroup label="Paternal grandparents">
-          <option value="paternal_grandmother">Paternal grandmother</option>
-          <option value="paternal_grandfather">Paternal grandfather</option>
-        </optgroup>
-      </select>
-      <div class="ctx-autocomplete-wrapper ctx-family-condition-wrap">
-        <input type="text" class="ctx-note-input" id="fh-condition" placeholder="Condition (e.g. heart attack, Alzheimer's, breast cancer)" oninput="filterFamilyConditionSuggestions()" onfocus="filterFamilyConditionSuggestions()" aria-label="Condition">
-        <div class="ctx-suggestions" id="fh-condition-suggestions"></div>
-      </div>
-    </div>
-    <div class="ctx-family-add-row">
-      <input type="number" min="0" max="120" class="ctx-note-input ctx-family-age-input" id="fh-age" placeholder="Age at onset" aria-label="Age at onset">
-      <input type="text" class="ctx-note-input ctx-family-note-input" id="fh-note" placeholder="Note — outcome, treatment, etc. (optional)" aria-label="Note">
-      <button class="import-btn import-btn-primary" onclick="addFamilyHistoryEntry()">+ Add</button>
-    </div>
-  </div></div>`;
-  html += renderNoteField(current.note);
-  const hasCurrent = conditions.length > 0 || familyHistory.length > 0 || current.note;
-  html += `<div class="ctx-editor-actions">
-    <button class="import-btn import-btn-primary" onclick="saveDiagnoses()">Save</button>
-    <button class="import-btn import-btn-secondary" onclick="closeDiagnoses()">Cancel</button>
-    ${hasCurrent ? '<button class="import-btn import-btn-secondary" style="color:var(--red);border-color:var(--red);margin-left:auto" onclick="clearDiagnoses()">Clear</button>' : ''}
-  </div>`;
-  renderContextEditorModal(modal, 'Medical History', 'Your diagnoses and family history. The AI considers both when interpreting your labs.', html, 'closeDiagnoses');
-  setTimeout(() => {
-    const input = document.getElementById('condition-input');
-    if (input) {
-      input.onkeydown = (e) => { if (e.key === 'Enter') { e.preventDefault(); addCondition(); } };
-    }
-    const fhCond = document.getElementById('fh-condition');
-    if (fhCond) {
-      fhCond.onkeydown = (e) => { if (e.key === 'Enter') { e.preventDefault(); addFamilyHistoryEntry(); } };
-    }
-    document.removeEventListener('click', closeSuggestionsOnClickOutside);
-    document.addEventListener('click', closeSuggestionsOnClickOutside);
-  }, 50);
-}
-
-export function filterConditionSuggestions() {
-  const input = document.getElementById('condition-input');
-  const container = document.getElementById('condition-suggestions');
-  if (!input || !container) return;
-  const val = input.value.toLowerCase().trim();
-  const existing = (state.importedData.diagnoses && state.importedData.diagnoses.conditions || []).map(c => c.name.toLowerCase());
-  const sexFiltered = COMMON_CONDITIONS.filter(c => {
-    if (state.profileSex === 'male' && (c === 'PCOS' || c === 'Endometriosis')) return false;
-    return true;
-  });
-  const matches = val ? sexFiltered.filter(c => c.toLowerCase().includes(val) && !existing.includes(c.toLowerCase())) : sexFiltered.filter(c => !existing.includes(c.toLowerCase()));
-  if (matches.length === 0 || !val) { container.innerHTML = ''; return; }
-  // JSON.stringify + escapeHTML so conditions with apostrophes (Alzheimer's,
-  // Hashimoto's, Crohn's, etc.) survive both the JS-string-in-HTML-attribute
-  // round-trip. escapeHTML alone would convert `'` → `&#39;` which the HTML
-  // parser then decodes *before* JS sees it, breaking the JS string literal.
-  container.innerHTML = matches.slice(0, 8).map(m => `<div class="ctx-suggestion-item" onmousedown="selectConditionSuggestion(${escapeHTML(JSON.stringify(m))})">${escapeHTML(m)}</div>`).join('');
-}
-
-export function selectConditionSuggestion(name) {
-  const input = document.getElementById('condition-input');
-  if (input) input.value = name;
-  const container = document.getElementById('condition-suggestions');
-  if (container) container.innerHTML = '';
-}
-
-export function closeSuggestionsOnClickOutside(e) {
-  const container = document.getElementById('condition-suggestions');
-  const input = document.getElementById('condition-input');
-  if (container && input && !input.contains(e.target) && !container.contains(e.target)) {
-    container.innerHTML = '';
-  }
-  // Family-history input shares this dismiss-on-outside-click behavior.
-  const fhContainer = document.getElementById('fh-condition-suggestions');
-  const fhInput = document.getElementById('fh-condition');
-  if (fhContainer && fhInput && !fhInput.contains(e.target) && !fhContainer.contains(e.target)) {
-    fhContainer.innerHTML = '';
-  }
-}
-
-export function syncDiagnosesNote() {
-  const noteEl = document.getElementById('ctx-note-input');
-  if (noteEl && state.importedData.diagnoses) state.importedData.diagnoses.note = noteEl.value.trim();
-}
-
-export function addCondition() {
-  const input = document.getElementById('condition-input');
-  const severity = getSelectedOption('condition-severity') || 'mild';
-  const since = document.getElementById('condition-since');
-  const name = input ? input.value.trim() : '';
-  if (!name) return;
-  syncDiagnosesNote();
-  if (!state.importedData.diagnoses) state.importedData.diagnoses = { conditions: [], note: '' };
-  const cond = { name, severity };
-  if (since && since.value.trim()) cond.since = since.value.trim();
-  state.importedData.diagnoses.conditions.push(cond);
-  recordChange('diagnoses');
-  saveImportedData();
-  renderDiagnosesModal(document.getElementById("detail-modal"), state.importedData.diagnoses);
-}
-
-export function deleteCondition(idx) {
-  if (!state.importedData.diagnoses || !state.importedData.diagnoses.conditions) return;
-  syncDiagnosesNote();
-  state.importedData.diagnoses.conditions.splice(idx, 1);
-  recordChange('diagnoses');
-  saveImportedData();
-  renderDiagnosesModal(document.getElementById("detail-modal"), state.importedData.diagnoses);
-}
-
-export function addFamilyHistoryEntry() {
-  const relativeEl = document.getElementById('fh-relative');
-  const conditionEl = document.getElementById('fh-condition');
-  const ageEl = document.getElementById('fh-age');
-  const noteEl = document.getElementById('fh-note');
-  const relative = relativeEl?.value || '';
-  const condition = (conditionEl?.value || '').trim();
-  if (!relative || !condition) return;
-  // Validate relative against allowlist so a tampered <option> can't sneak in.
-  if (!FAMILY_RELATIVES.some(r => r.key === relative)) return;
-  const ageRaw = (ageEl?.value || '').trim();
-  const onsetAge = ageRaw === '' ? null : Math.max(0, Math.min(120, parseInt(ageRaw, 10)));
-  const note = (noteEl?.value || '').trim();
-  syncDiagnosesNote();
-  if (!state.importedData.diagnoses) state.importedData.diagnoses = { conditions: [], note: '', familyHistory: [] };
-  if (!Array.isArray(state.importedData.diagnoses.familyHistory)) state.importedData.diagnoses.familyHistory = [];
-  const entry = { relative, condition };
-  if (Number.isFinite(onsetAge)) entry.onsetAge = onsetAge;
-  if (note) entry.note = note;
-  state.importedData.diagnoses.familyHistory.push(entry);
-  recordChange('diagnoses');
-  saveImportedData();
-  renderDiagnosesModal(document.getElementById("detail-modal"), state.importedData.diagnoses);
-}
-
-export function deleteFamilyHistoryEntry(idx) {
-  if (!state.importedData.diagnoses || !Array.isArray(state.importedData.diagnoses.familyHistory)) return;
-  syncDiagnosesNote();
-  state.importedData.diagnoses.familyHistory.splice(idx, 1);
-  recordChange('diagnoses');
-  saveImportedData();
-  renderDiagnosesModal(document.getElementById("detail-modal"), state.importedData.diagnoses);
-}
-
-// Mirrors filterConditionSuggestions but targets the family-history input
-// — same COMMON_CONDITIONS pool, separate DOM ids so the two autocompletes
-// don't share state.
-export function filterFamilyConditionSuggestions() {
-  const input = document.getElementById('fh-condition');
-  const container = document.getElementById('fh-condition-suggestions');
-  if (!input || !container) return;
-  const val = input.value.toLowerCase().trim();
-  const sexFiltered = COMMON_CONDITIONS.filter(c => {
-    // Cross-sex family members can have either-sex conditions, so don't apply
-    // the user's profile sex filter here — a male profile can have a mother
-    // with PCOS or endometriosis.
-    return true;
-  });
-  const matches = val ? sexFiltered.filter(c => c.toLowerCase().includes(val)) : sexFiltered;
-  if (matches.length === 0 || !val) { container.innerHTML = ''; return; }
-  // Same JSON.stringify-then-escapeHTML trick as filterConditionSuggestions —
-  // see the comment there for why bare escapeHTML breaks apostrophe names.
-  container.innerHTML = matches.slice(0, 8).map(m => `<div class="ctx-suggestion-item" onmousedown="selectFamilyConditionSuggestion(${escapeHTML(JSON.stringify(m))})">${escapeHTML(m)}</div>`).join('');
-}
-
-export function selectFamilyConditionSuggestion(name) {
-  const input = document.getElementById('fh-condition');
-  if (input) input.value = name;
-  const container = document.getElementById('fh-condition-suggestions');
-  if (container) container.innerHTML = '';
-}
-
-export function saveDiagnoses() {
-  const note = (document.getElementById('ctx-note-input') || {}).value || '';
-  if (!state.importedData.diagnoses) state.importedData.diagnoses = { conditions: [], note: '' };
-  state.importedData.diagnoses.note = note.trim();
-  const fhLen = Array.isArray(state.importedData.diagnoses.familyHistory) ? state.importedData.diagnoses.familyHistory.length : 0;
-  // Don't nuke the whole card if family history is the only populated field.
-  if (state.importedData.diagnoses.conditions.length === 0 && !state.importedData.diagnoses.note && fhLen === 0) {
-    state.importedData.diagnoses = null;
-  }
-  saveAndRefresh('Medical history saved', 'diagnoses');
-}
-
-export function closeDiagnoses() {
-  window.closeModal();
-}
-
-export function clearDiagnoses() {
-  state.importedData.diagnoses = null;
-  saveAndRefresh('Medical history cleared', 'diagnoses');
 }
 
 // ═══════════════════════════════════════════════
