+++ service-worker.js
@@ -102,6 +102,7 @@ const APP_SHELL = [
   '/js/context-cards.js',
   '/js/context-card-summaries.js',
   '/js/context-card-editor-ui.js',
+  '/js/context-card-medical-history-editor.js',
   '/js/focus-card.js',
   '/js/onboarding-view.js',
   '/js/emf-facade.js',
